refactor: deepen delegated child run supervision
